// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/waist-history-page-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDesiredWaistHttpResponse } from './user-desired-waist-http-response';
import { WaistEntryHttpResponse } from './waist-entry-http-response';
import { WaistEntrySummaryHttpResponse } from './waist-entry-summary-http-response';
import { WaistGoalHistoryHttpResponse } from './waist-goal-history-http-response';

export interface WaistHistoryPageSummaryHttpResponse {
    entries?: Array<WaistEntryHttpResponse>;
    goal?: UserDesiredWaistHttpResponse;
    goalHistory?: Array<WaistGoalHistoryHttpResponse>;
    heightCm?: number | null;
    summary?: Array<WaistEntrySummaryHttpResponse>;
}