// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/weight-history-page-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDesiredWeightHttpResponse } from './user-desired-weight-http-response';
import { WeightEntryHttpResponse } from './weight-entry-http-response';
import { WeightEntrySummaryHttpResponse } from './weight-entry-summary-http-response';
import { WeightGoalHistoryHttpResponse } from './weight-goal-history-http-response';

export interface WeightHistoryPageSummaryHttpResponse {
    entries?: Array<WeightEntryHttpResponse>;
    goal?: UserDesiredWeightHttpResponse;
    goalHistory?: Array<WeightGoalHistoryHttpResponse>;
    heightCm?: number | null;
    summary?: Array<WeightEntrySummaryHttpResponse>;
}